requirements.txt
torch>=1.10
ffmpeg-python
soundfile
tqdm
numpy
Pillow
